+++ CHANGELOG.md
@@ -13,6 +13,7 @@ mod).
 
 ### Changed
 
+- Updated Mash, Thicken and Twist to preview the Mix token they would create instead of using the hover tips
 - Updated some Uncommon and Rare cards to use the cast animation
 
 ## [0.14.0] - 2026-08-29